requirements.txt
# Core Flask dependencies
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.1.0

# Document processing
python-docx>=1.1.0
docx2txt>=0.8

# Scientific computing and NLP
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
nltk>=3.8

# Data handling
pandas>=2.0.0

# Web server
gunicorn>=21.0.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
lxml>=4.9.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
wordcloud>=1.9.0
